chore: move route components into views
