+++ src/App.vue
@@ -10,10 +10,10 @@
 </template>
 
 <script lang="ts">
+import { useDark, useToggle } from '@vueuse/core';
 import { Component, Vue } from 'vue-property-decorator';
 import Cookies from 'js-cookie';
 import NavBar from '@/components/NavBar/NavBar.vue';
-import { useDark, useToggle } from '@vueuse/core';
 
 const isDark = useDark({
   attribute: 'color-scheme',
@@ -29,7 +29,7 @@ export default class App extends Vue {
   public loggedIn = false;
   public darkMode = isDark.value;
 
-  public switchTheme() {
+  protected switchTheme() {
     toggleDark();
     this.darkMode = !this.darkMode;
   }
@@ -42,8 +42,6 @@ export default class App extends Vue {
     } else {
       this.loggedIn = false;
     }
-
-    Cookies.set('pulse-light-theme', String(isDark.value));
   }
 }
 </script>
