+++ src/views/user/UserDetails.vue
